reduced db writes

+++ backend/src/db/index.js
@@ -1,7 +1,6 @@
 const fs = require('fs');
 const path = require('path');
 
-
 const TEE_FOLDER_DB_NAME = `tee_db.json`;
 
 let db = null;
@@ -10,18 +9,39 @@ class Db {
     constructor(filePath) {
         this.filePath = filePath;
         this.data = {};
+        this.dirty = true;
         if (fs.existsSync(this.filePath)) {
             this.data = JSON.parse(fs.readFileSync(this.filePath));
         }
     }
 
-    write() {
-        fs.writeFileSync(this.filePath, JSON.stringify(this.data));
+    save(data) {
+        this.data = {
+            ...this.data,
+            ...data
+        };
+        this.dirty = true;
+    }
+
+    /**
+     * Writes changes in-memory 
+     * @param {boolean} force - if true will force writing 
+     * */
+    write(force = false) {
+        if (this.dirty || force) {
+            fs.writeFileSync(this.filePath, JSON.stringify(this.data));
+            this.dirty = false;
+            return true;
+        }
+
+        return false;
     }
 
     delete() {
         if (fs.existsSync(this.filePath)) {
             fs.rmSync(this.filePath);
+            this.data = {};
+            this.dirty = false;
         }
     }
 }
@@ -31,6 +51,7 @@ const init = filePath => {
 };
 
 module.exports = {
+    /** @return {Db} */
     getDb(filePath = null) {
         if (!Boolean(db) && !Boolean(filePath)) throw Error(`Cannot init db without filepath, passed ${filePath}`);
